Ensure `javalib` source compatibility for Scala 3 build (#2457)
* Move java classes registered in Scala 3 compiler to scala-2 sources. Create renamed versions for Scala 3

* Seperate implementation of Scala 2/3 enums

* Rewrite do-while into while-do loops

* Remove `final` modifiler from object

* Fix not supported matches with @swtich annotation

* Fix incorrect signatures of class members (replace val foo with def foo(), etc)

* Fix type mismatch errors

* Fix type bounds for not incorrect method signatures

* Address deprecated methods in j.l.reflect.Array

* Add parens around lambda parameters, add explicit type for pointers

* Add ScalaStream compat for class

* Fix complation problems with java.nio buffers

* Duplicate java.lang.ProcessBuilder for scala 2 and 3, due to inner definition of enums

* Split java.util.Formatter into Impl class to handle problem with inner enum defintion

* Adapt scalafmt to handle Scala3 sourcesand run scalafmt

* Add testing javalib and scalalib compilation in the CI

* Fix various typos

* Add missing explenations

* Split ProcessBuilder into Impl and actual class

* Run scalafmt

+++ javalib/src/main/scala/java/io/RandomAccessFile.scala
@@ -154,7 +154,7 @@ class RandomAccessFile private (
       val fileLength = length()
       val toSkip =
         if (currentPosition + n > fileLength) fileLength - currentPosition
-        else n
+        else n.toLong
       seek(toSkip)
       toSkip.toInt
     }
